requirements.txt
accelerate==0.33.0
datasets==2.21.0
easydict==1.13
einops==0.6.1
fairscale==0.4.13
fastapi==0.112.2
ftfy==6.1.3
h5py==3.14.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.24.4
openai==1.43.0
opencv-contrib-python-headless==4.6.0.66
opencv-python-headless==4.6.0.66
PyYAML==6.0.1
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.11.4
tensorboard==2.17.1
tensorboard-data-server==0.7.2
timm==1.0.3
tokenizers==0.19.1
torchdiffeq==0.2.3
tqdm==4.66.5
transformers==4.43.0
triton==2.2.0
sentencepiece==0.1.99
decord==0.6.0
